Using settings factory and adding CONTRIBUTING file. https://github.com/ORNL/flowcept/issues/27
