+++ flowcept/flowceptor/plugins/mlflow/mlflow_interceptor.py
@@ -2,8 +2,8 @@ import sys
 import os
 import time
 from watchdog.observers import Observer
-from flowcept.flowceptor.plugins.abstract_flowceptor import (
-    AbstractFlowceptor,
+from flowcept.flowceptor.plugins.base_interceptor import (
+    BaseInterceptor,
 )
 from flowcept.flowceptor.plugins.interceptor_state_manager import (
     InterceptorStateManager,
@@ -15,7 +15,7 @@ from flowcept.flowceptor.plugins.mlflow.interception_event_handler import (
 )
 
 
-class MLFlowInterceptor(AbstractFlowceptor):
+class MLFlowInterceptor(BaseInterceptor):
     def __init__(self, plugin_key="mlflow"):
         super().__init__(plugin_key)
         self.state_manager = InterceptorStateManager(self.settings)
